Ajout du traitement de la session et la déconnexion

+++ src/java/controller/Controlleur.java
@@ -10,6 +10,7 @@ import java.sql.SQLException;
 import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import model.Customer;
 import utils.MagasinHelper;
 import org.springframework.web.servlet.ModelAndView;
@@ -27,6 +28,10 @@ public class Controlleur extends MultiActionController {
     public ModelAndView afficherClients(HttpServletRequest request,
             HttpServletResponse response) throws Exception {
         try {
+            HttpSession session = request.getSession(false);
+            if (session.getAttribute("nameCurrentUser") == null) {
+                return new ModelAndView("login");
+            }
             requeteur = new MagasinHelper();
             resultatrequete a = new resultatrequete();
             a.setResult(requeteur.getClients());
@@ -39,6 +44,10 @@ public class Controlleur extends MultiActionController {
     public ModelAndView detailClient(HttpServletRequest request,
             HttpServletResponse response) throws Exception {
         try {
+            HttpSession session = request.getSession(false);
+            if (session.getAttribute("nameCurrentUser") == null) {
+                return new ModelAndView("login");
+            }
             requeteur = new MagasinHelper();
             resultatrequete a = new resultatrequete();
             a.setClient(requeteur.getClientById(Integer.parseInt(request.getParameter("Operation"))));
@@ -59,6 +68,10 @@ public class Controlleur extends MultiActionController {
     public ModelAndView afficherFormInscriptionClient(HttpServletRequest request,
             HttpServletResponse response) throws Exception {
         try {
+            HttpSession session = request.getSession(false);
+            if (session.getAttribute("nameCurrentUser") == null) {
+                return new ModelAndView("login");
+            }
             requeteur = new MagasinHelper();
             resultatrequete a = new resultatrequete();
             resultatrequete b = new resultatrequete();
@@ -78,6 +91,10 @@ public class Controlleur extends MultiActionController {
     public ModelAndView enregistrerClient(HttpServletRequest request,
             HttpServletResponse response) throws Exception {
         try {
+            HttpSession session = request.getSession(false);
+            if (session.getAttribute("nameCurrentUser") == null) {
+                return new ModelAndView("login");
+            }
             requeteur = new MagasinHelper();
             String param1 = request.getParameter("nom");
             String param2 = request.getParameter("adresse");
@@ -97,6 +114,10 @@ public class Controlleur extends MultiActionController {
     public ModelAndView modifierClient(HttpServletRequest request,
             HttpServletResponse response) throws Exception {
         try {
+            HttpSession session = request.getSession(false);
+            if (session.getAttribute("nameCurrentUser") == null) {
+                return new ModelAndView("login");
+            }
             requeteur = new MagasinHelper();
             String param1 = request.getParameter("numero");
             String param2 = request.getParameter("nom");
@@ -117,6 +138,10 @@ public class Controlleur extends MultiActionController {
     public ModelAndView supprimerClient(HttpServletRequest request,
             HttpServletResponse response) throws Exception {
         try {
+            HttpSession session = request.getSession(false);
+            if (session.getAttribute("nameCurrentUser") == null) {
+                return new ModelAndView("login");
+            }
             requeteur = new MagasinHelper();
             String param1 = request.getParameter("numero");
             requeteur.deleteCustomer(Integer.parseInt(param1));
@@ -131,6 +156,10 @@ public class Controlleur extends MultiActionController {
     public ModelAndView afficherAchats(HttpServletRequest request,
             HttpServletResponse response) throws Exception {
         try {
+            HttpSession session = request.getSession(false);
+            if (session.getAttribute("nameCurrentUser") == null) {
+                return new ModelAndView("login");
+            }
             requeteur = new MagasinHelper();
             resultatrequete a = new resultatrequete();
             a.setResult(requeteur.getAchats(Integer.parseInt(request.getParameter("numero"))));
@@ -146,13 +175,19 @@ public class Controlleur extends MultiActionController {
     public ModelAndView home(HttpServletRequest request,
             HttpServletResponse response) throws Exception {
 
+        HttpSession session = request.getSession(false);
+        if (session.getAttribute("nameCurrentUser") == null) {
+            session.setAttribute("nameCurrentUser", request.getUserPrincipal().toString());
+        }
         return new ModelAndView("home");
-
     }
 
     public ModelAndView afficherFormRechercheClient(HttpServletRequest request,
             HttpServletResponse response) throws Exception {
-
+        HttpSession session = request.getSession(false);
+        if (session.getAttribute("nameCurrentUser") == null) {
+            return new ModelAndView("login");
+        }
         return new ModelAndView("recherche");
     }
 
@@ -160,6 +195,10 @@ public class Controlleur extends MultiActionController {
             HttpServletResponse response) throws Exception {
 
         try {
+            HttpSession session = request.getSession(false);
+            if (session.getAttribute("nameCurrentUser") == null) {
+                return new ModelAndView("login");
+            }
             String identifiant_client = request.getParameter("identifiant_client");
             String nom_client = request.getParameter("nom_client");
             if (!identifiant_client.isEmpty()) {
@@ -213,4 +252,13 @@ public class Controlleur extends MultiActionController {
 
     }
 
+    public ModelAndView disconnect(HttpServletRequest request,
+            HttpServletResponse response) throws Exception {
+
+        HttpSession session = request.getSession(false);
+        if (session != null) {
+            session.invalidate();
+        }
+        return new ModelAndView("login");
+    }
 }
